mpc85xx: convert TL-WDR4900 v1 to dtbImage
Converts the TP-Link WDR4900 v1 to use the dtbImage in the
hopes of prolonging the life of the device. While at it,
the patch makes the fdt.bin an ARTIFACT and sets the KERNEL_SIZE
to 2684 KiB as a precaution since the stock u-boot is using a
fixed kernel size.

Note: Give the image some time, it will take much longer to
extract and boot.

[compiled tested for 4.14]