x86: fix geode image builds redux
This patch adds the disabled DRM_RADEON and DRM_AMDGPU
config symbols from the x86' config to the generic target
configs. The existing symbols in the x86' configs are kept
for now, until we know whenever we want to remove such
symbols or not (see Github PR #1831, #1825, #1828).

And hopefully this time for 4.14 and 4.19.

Signed-off-by: Christian Lamparter <chunkeey@gmail.com>